backend/requirements.txt
aiofiles==24.1.0
alembic==1.13.3
anthropic==0.40.0
asyncpg==0.30.0
bcrypt==4.2.1
black==24.10.0
fastapi==0.115.0
flake8==7.1.1
httpx==0.27.2
langchain==0.3.7
langchain-anthropic==0.2.4
langchain-community==0.3.5
matplotlib==3.9.2
mangum==0.17.0
mypy==1.13.0
numpy==2.1.3
openai==1.54.0
openpyxl==3.1.5
pandas==2.2.3
passlib[bcrypt]==1.7.4
pgvector==0.3.0
plotly==5.24.1
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.0
PyPDF2==3.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
seaborn==0.13.2
sentence-transformers==3.3.1
sqlalchemy==2.0.35
uvicorn[standard]==0.32.0
